requirements.txt
streamlit==1.29.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
plotly==5.18.0
gspread==5.12.3
oauth2client==4.1.3
python-dotenv==1.0.0
openpyxl==3.1.2
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
biopython==1.83
pyyaml==6.0.1
webdriver-manager==4.0.1
lxml==4.9.3
fake-useragent==1.4.0
